Fix: can't play next song automatically

+++ src/core/music/MusicPlayer.ts
@@ -1,5 +1,4 @@
 import {
-    Client,
     CommandInteraction,
     Guild,
     GuildMember,
@@ -18,9 +17,11 @@ import {
     VoiceConnection,
     AudioPlayer,
     AudioPlayerError,
+    NoSubscriberBehavior,
 } from "@discordjs/voice";
 import { Logger } from "../utils/Logger.js";
 import { stream, video_info, YouTubeVideo } from "play-dl";
+import { promisify } from "node:util";
 
 interface Metadata {
     videoInfo: YouTubeVideo;
@@ -33,7 +34,6 @@ interface Track {
 }
 
 export class MusicPlayer {
-    private client: Client;
     private guild: Guild;
     private channel: TextBasedChannel;
     private queue: Track[];
@@ -47,7 +47,6 @@ export class MusicPlayer {
     public destroyed: boolean;
 
     constructor(interaction: CommandInteraction) {
-        this.client = interaction.client;
         this.guild = interaction.guild;
         this.channel = interaction.channel;
 
@@ -76,12 +75,12 @@ export class MusicPlayer {
                     newState.closeCode === 4014
                 ) {
                     /*
-						If the WebSocket closed with a 4014 code, this means that we should not manually attempt to reconnect,
-						but there is a chance the connection will recover itself if the reason of the disconnect was due to
-						switching voice channels. This is also the same code for the bot being kicked from the voice channel,
-						so we allow 5 seconds to figure out which scenario it is. If the bot has been kicked, we should destroy
-						the voice connection.
-					*/
+        				If the WebSocket closed with a 4014 code, this means that we should not manually attempt to reconnect,
+        				but there is a chance the connection will recover itself if the reason of the disconnect was due to
+        				switching voice channels. This is also the same code for the bot being kicked from the voice channel,
+        				so we allow 5 seconds to figure out which scenario it is. If the bot has been kicked, we should destroy
+        				the voice connection.
+        			*/
                     try {
                         await entersState(
                             this.connection,
@@ -95,19 +94,22 @@ export class MusicPlayer {
                     }
                 } else if (this.connection.rejoinAttempts < 5) {
                     /*
-						The disconnect in this case is recoverable, and we also have <5 repeated attempts so we will reconnect.
-					*/
-                    setTimeout(() => this.connection.rejoin(), 5000);
+        				The disconnect in this case is recoverable, and we also have <5 repeated attempts so we will reconnect.
+        			*/
+                    await promisify(setTimeout)(
+                        (this.connection.rejoinAttempts + 1) * 5000
+                    );
+                    this.connection.rejoin();
                 } else {
                     /*
-						The disconnect in this case may be recoverable, but we have no more remaining attempts - destroy.
-					*/
+        				The disconnect in this case may be recoverable, but we have no more remaining attempts - destroy.
+        			*/
                     this.destroy();
                 }
             } else if (newState.status === VoiceConnectionStatus.Destroyed) {
                 /*
-					Once destroyed, stop the subscription
-				*/
+        			Once destroyed, stop the subscription
+        		*/
                 this.destroy();
             } else if (
                 !this.readyLock &&
@@ -115,10 +117,10 @@ export class MusicPlayer {
                     newState.status === VoiceConnectionStatus.Signalling)
             ) {
                 /*
-					In the Signalling or Connecting states, we set a 20 second time limit for the connection to become ready
-					before destroying the voice connection. This stops the voice connection permanently existing in one of these
-					states.
-				*/
+        			In the Signalling or Connecting states, we set a 20 second time limit for the connection to become ready
+        			before destroying the voice connection. This stops the voice connection permanently existing in one of these
+        			states.
+        		*/
                 this.readyLock = true;
                 try {
                     await entersState(
@@ -126,20 +128,21 @@ export class MusicPlayer {
                         VoiceConnectionStatus.Ready,
                         20000
                     );
-                    this.readyLock = false;
                 } catch {
                     if (
                         this.connection.state.status !==
                         VoiceConnectionStatus.Destroyed
                     )
                         this.destroy();
+                } finally {
+                    this.readyLock = false;
                 }
             }
         });
 
         this.player = createAudioPlayer({
             behaviors: {
-                maxMissedFrames: 99999999,
+                noSubscriber: NoSubscriberBehavior.Play,
             },
             debug: Boolean(process.env.DEBUG),
         });
@@ -147,14 +150,19 @@ export class MusicPlayer {
 
         // Configure audio player
         this.player.on("stateChange", async (oldState, newState) => {
+            Logger.debug(
+                `State Changed: from ${oldState.status} to ${newState.status}`
+            );
             if (
                 newState.status === AudioPlayerStatus.Idle &&
                 oldState.status !== AudioPlayerStatus.Idle
             ) {
                 // If the Idle state is entered from a non-Idle state, it means that an audio resource has finished playing.
                 // The queue is then processed to start playing the next track, if one is available.
-                if (!this.np.deleted && this.np.deletable)
-                    await this.np.delete();
+                Logger.debug(
+                    `${this.current.metadata.videoInfo.title} finished playing`
+                );
+                if (this.np.deletable) await this.np.delete();
                 this.np = null;
                 this.current = null;
                 this.processQueue();
@@ -168,8 +176,7 @@ export class MusicPlayer {
                 `${error.name}: ${error.message} with resource ${
                     (error.resource as AudioResource<Metadata>).metadata
                         .videoInfo.title
-                }
-                Stack: ${error.stack}`
+                }`
             );
             await this.channel.send("嗯....似乎沒辦法唱下去的樣子...");
         });
@@ -223,8 +230,14 @@ export class MusicPlayer {
      * Attempts to play a Track from the queue
      */
     private async processQueue() {
+        Logger.debug(`Processing queue of ${this.guild.name}`);
+        if (this.destroyed) {
+            Logger.warn(
+                `Music player of ${this.guild.name} has already been destroyed`
+            );
+            return;
+        }
         if (
-            this.destroyed ||
             this.queueLock ||
             this.player.state.status !== AudioPlayerStatus.Idle ||
             this.queue.length === 0
@@ -235,8 +248,13 @@ export class MusicPlayer {
                 );
                 this.destroy();
             }
+            Logger.debug(
+                `Queue of ${this.guild.name} locked or player is busy, skipping`
+            );
             return;
         }
+
+        Logger.debug("Preparing next song");
 
         this.queueLock = true;
 
@@ -253,7 +271,11 @@ export class MusicPlayer {
             );
             this.player.play(resource);
             this.queueLock = false;
+            Logger.debug(
+                `${this.current.metadata.videoInfo.title} started playing`
+            );
         } catch (err) {
+            Logger.error("Error occurs when preparing next song", err);
             this.queueLock = false;
             return this.processQueue();
         }
@@ -263,11 +285,15 @@ export class MusicPlayer {
      * Disconnect and cleanup the player.
      */
     public destroy() {
-        if (this.connection.state.status !== VoiceConnectionStatus.Destroyed)
+        Logger.info("Destroying player");
+        if (this.connection.state.status !== VoiceConnectionStatus.Destroyed) {
             this.connection.destroy();
+            Logger.info("Voice connection destroyed");
+        }
         this.player.stop(true);
         this.queue = null;
         this.destroyed = true;
+        Logger.info("Player destroyed");
     }
     /**
      * Change connected voice channel
